// src/lib/api/generated/model/status37cEnum.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Django Continuous Integration API
 * API documentation for the Django Continuous Integration project with tenant-aware access rules.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `DRAFT` - Draft
* `ACTIVE` - Active
* `ARCHIVED` - Archived
 */
export type Status37cEnum = typeof Status37cEnum[keyof typeof Status37cEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const Status37cEnum = {
  DRAFT: 'DRAFT',
  ACTIVE: 'ACTIVE',
  ARCHIVED: 'ARCHIVED',
} as const;